Updating Code
I updated much of the code, having made the text based interface, and have started to make progress on the graphs that will be visualised. Parts that I need to improve are issues with getting VS to read the .csv file, and some general problems with the code that I have typed out wrong.